Move onewire entity classes to separate module (#42445)
